Texcraft Playground v1
This commit contains the inital unpolished version of the Texcraft Playground.
This is a website for running TeX scripts, based on Texcraft, and hosted at
play.texcraft.dev. It's focused on TeX as a language, and in v1 only supports
running the "Turing complete subset" of TeX. Probably it will stay this way
forever: actually outputting documents in the browser is a whole other story.

The stack is WASM/TypeScript/Go and right now is being put under the
playground directory, but this may change when the repo moves to a multi crate
architecture.

I mentioned it's unpolished: it's really unpolished. There are a bunch of
TODOs in the code and in playground/README.md. Resolving this will involve
work in the core Texcraft code as well as the new playground code.