Digits of pi TeX script
Adds a script to calculate the first n digits of pi. This script has been added
to the Playground. More importantly, a version of the script has been added
to a new performance directory and is the first benchmark of TeX. Initial
perf results are pretty bad! Texcraft runs the script seven times slower than
pdfTeX.

Adding this script necessitated implementing the \let command and making
the number of \count registers configurable.