Refactor TeX macro code into two modules (#1)
This commit splits the TeX macro code in two. The macro data structures and
the code for expanding macros are put in a new macrotypes modules. The code
for creating a macro using the \def primitive stays in the def module. This
layout better reflects the structure of the TeX macros system, and helps keep
cognitive load down.

The main motivation, however, was to make the upcoming optimizations of
macros easier. The Texlang runtime will more closely integrate with the macro
expansion logic so it makes sense to take it out of the Texlang stdlib.

On the way a first version of the \tracingmacros command was implemented.