Simply the input controller
This commit is mainly about simplifying the code in the input controller. This
results in a performance gain becasue all of the code in the controller is
really hot.

At the same time, based on my increased understanding of the side effects of
token peeking in TeX [1], I ripped out the immutable peek code. In fact, it was
unused and was making Texcraft more complicated and actually slower.

[1] https://jpfennell.com/posts/tex-peek-cat-codes