String interning for control sequence names
This commit makes two changes to the token type:

- The 24 byte small string type used for control sequence names is replaced by
  an interned string type that is 4 bytes.

- The source field used for tracebacks is temporarily removed.

With these two changes the size of the token type goes down from ~50 bytes
to 12 bytes. This results in a 33% speed up of all benchmarks. The
advance benchmark went from 500ns to 330ns, and similarly for the digits of
pi benchmark.

Funnily enough, Knuth's TeX also uses a string interning system for control
sequence names. Ours has one advantage though: if the same string is interned
twice the ID returned is the same. I believe this is *not* true of Knuth's
TeX. This feature means we can compare strings using a single u32 comparison.
It also opens up the possibility, to be explored, of replacing the commands
map with a commands array, keyed off of the string interning key.