Input and macro optimizations
This is a refactor of two key sytems in Texcraft: the input controller and
the macro expansion system. The refactors were motivated by performance and
together have improved performance numbers again by a non-trivial percent
(maybe 10 to 20 percent).

The input controller changes are guided by the idea that the happy path
of getting the next valid token (either from an expansion or from lexing a
file) should be prioritized and error handling should happen last. The numbers
validate the approach. On the other hand, the advance benchmark previously
did a lot of non-happy-path work of adding new sources, so that benchmark
increased. It was changed to again just exercise the happy path.

The macro changes are pretty common-sense and rely on tokens now being
copiable. In particular, extending a vector of one tokens with another vector
can be optimized down to a memcpy, and the refactor enables this opimization.

As an aside, the commit adds a way to customize the n in the digits of pi
benchmark using an env var to `cargo bench`.