Driver and input rearchitecture
This commit merges together the driver and input Rust modules and more
tightly integrates the input controller with the execution input type. The
motivation is for the code to be simpler, and to make further improvements
easier.

At the same time, the execution input is changed to own the input controller
and the state. This makes it a little more awkward to drive TeX execution,
but makes the code simpler by getting rid of the lifetimes on the execution
input type.

Performance change is negligible.