Change the structure of the token type
With this commit the structure of the token type is changed, mostly
motivated by reducing its size. With this change, the token value type is now
8 bytes instead of 12, and this gives a small expected performance gain.

Before, token value was a 2-valued enum with character and control sequence
variants, and the character variant included the character itself and the
cat code. There are 11 admissible cat codes. With this change, the token
value now has 12 variants corresponding to control sequence plus all 11
possible character codes. Each variant includes a utf8 character, except
for the control sequence which contains a string, of course. Given how Rust
implements enums its clear where the 12 to 8 byte size reduction comes from.

After this change it turned out the CatCode vs RawCatCode distinction was
not needed - in fact, it duplicated the variants in the token type. So
these were merged together.