Use a vector for storing commands
Any TeX implementation includes a map of control sequence names to commands.
In Texcraft, control sequence names are interned and then passed around
internally using a key that is basically an integer. Our string interner has
the property that the first string interned has key 1, the second key 2, and
so on. So the commands map has essentially integers as keys.

This commit changes the map from being based on HashMap to being based on
a plain old Vec. This change results in a 25 percent speedup in all benchmarks!
It has the con that if a control sequence name does not reference a command,
there will be holes in the vec that will take up space. In the long run, we
will need to see if this a problem and mitigate it. (Mitigation could mean
introducing a garbage collection system for csnames, not necessarily reverting
this commit). For the moment, we are happy with out speedup and the fact that
the digits of pi benchmark is only 44% slower than Knuth's TeX.

This commit also changes the opt-level from 3 to 2, and this results in
another speedup.