Texcraft and pdfTeX performance parity
This commit rewrites some of the driver code, especially around expansion,
to optimize for code paths that are most commonly taken (i.e., non expanding
code paths). This rewrite, along with small things like inlining, result in
a 40% performance improvement on the benchmarks, and finally (!) get Texcraft
to be as performant as pdfTeX.

The first use of unsafe was introduced to get around borrow checker
limitations, and some unsafe is used in the macro expansion code to avoid
a couple of redundant bounds checks.