Lexing architecture and performance improvements
This commit sort of re-architects the lexer to be a bit cleaner and to make
upcoming work on tracebacks easy to implement. The main change is reading
a source of TeX into a String upfront and then creating tokens from that.
In the tracebacks system, the String will be saved and used for printing lines
of source in error messages.

The structure of the cat code map is also changed to be based partly on a
vector rather than a map, for performance reasons. This is basically the same
optimization that was done for the commands map. With this change the lexer
throughput benchmark is cut in half, and is now below pdfTeX.