Refactor of the Texlang runtime code
This is a pretty dramatic refactor of the Texlang crate. The old state::Base
and InputController types are refactored into a new runtime::Env environment
that is the Texlang "runtime environment". This type cleanly separates:

1. State that is always needed in Texlang like the commands map.
2. Custom state.
3. Context related to expansion.
4. Internal parts of the environment like the CS name interner.

The old driver code then provides a layer on top of the Env that, via Rust's
visibility rules, restricts access to the Env except in allowed ways.