Add expansion state system
Some expansion commands need to maintain some state - currently, the
conditional commands do this. In general we don't want to give expansion
commands access to the state because it might break how we reason about TeX
parsing. The previous solution involved an ad hoc "expansion controller" that
had this state and was globally visible and mutable. This setup really cut
against Texlang's philosophy of minimizing state visibility as much as
possible.

This commit adds the ability to mark a part of the state as visible
to expansion commands using the HasExpansionState trait. It then migrates
the conditional commands to this new system.