Unify implementations of the component trait
The Texlang code currently uses a design pattern that, starting in this
commit, we call the component pattern. It's really nothing fancy but giving
it a name is nice. This pattern is really fundemental to Texlang and how it
improves on legacy TeX and giving it a name makes it easier to talk about.

More importantly, this commit adds some Rust formalism around implementing
this pattern. The key addition is a `HasComponent` trait that enforces that
a state type has the component needed for certain commands. All of the ad hoc
traits in the standard library have been replaces with this new trait.