Move the token stream trait into the runtime
In the initial Texcraft work the stream trait *seemed* like a great
abstraction. However it has turned out not to be useful in general. One reason
for this is that performance turned out to be harder than expected, and
removing dynamic trait patterns was important for solving this. Also the
streams turned out to be more coupled to the runtime environment than
anticipated, and having just a stream was rarely useful.

So, this commit moves the stream type into the runtime and starts coupling
it more with the environment.