Move the scratch space to the internal env.
This refactor is being done in preparation for an upcoming refactor in
which the ExpandedInput type will wrap the environment directly.

This commit causes the digits of pi benchmark to regress by about 3%. This
is because instead of having a static location for scratch space, we swap the
scratch space in and out. I think this 3% is worth it: otherwise, we may
risk memory safety issues. Ofc, it can always be revisited.

Also 5 usages of unsafe are removed in the macro code. These were premature
optimizations! The benchmarks didn't budge when they were removed.