Big refactor of the runtime
This is the latest commit that refactors the runtime, though I have a feeling
this may be the last. The thing that nagged me to refactor was that the run
function in the runtime accepted a `&mut ExecutionInput` which callers had to
construct themselves from an `Env`. Then, when finished running callers had to
extract the `Env` back out with a `take_env` method. This was clumsy: ideally
callers would just pass a `&mut Env`, though I couldn't figure out how to do
this.

The big breakthrough here was me learning about `#[repr(transparent)]` which
allows you to create a type that is identical to another type. It builds
on the newtype pattern. Adding the repr gives you something else though: it
allows you to cast from `&mut T` to `&mut S` *in both directions*. One of the
directions - the nested to the outer  - has to be done in an
unsafe block but it's actually safe.

For Texlang, I applied this pattern to `ExecutionInput` and all the other
input types. Now run can accept a `&mut Env` and we can convert it to a
`&mut ExecutionInput` and continue.

While in the runtime I also just moved code around a lot. The main thing
is the code now more cleanly constructs `ExecutionInput` et al as visibility
laters over the environment with the ability to read tokens. But all of the
token reading logic was extracted to a submodule.