Initial implementation of grouping semantics
In TeX, mutations made to variables are rolled back at the end of the group.
This commit implements this for Texlang. There is still some work to do, for
example supporting the execution commands \begingroup.

This commit also implement \global, which can be used to make a global mutation
to a variable. This prefix command works with arithemetic operators, \let,
\def and variables.