Token tracing system
This commit adds a system for tracing the source of a token using a 32-bit
key that is attached to it. The original tracing system in which data was
stored directly on the token using smart pointers was removed as part of the
performance work, and errors haven't worked since then. Errors are now working
again.

Information on the tracing system is described in the Rust module's docs.