Proof of concept of typed errors
So far in Texcraft all errors have been of type TokenError of EndOfInputError.
Effectively the specific error that occured is erased and encoded in the
string arguments of these errors. This makes the code a bit messy as lots
of space in functions is taken up constructing these kinds of errors. Also
it is impossible to programatically determine what specific error occured
without string inspection.

This commit adds a different error approach. In this approach, each module
gets its own error type, which is an enum with variants for each error. The
payload is enough to display the error nicely and typically includes a few
traces from the new tracing code. The formatting of this error is done by
manually implementing the Error trait. However, most of the work is done
by calling the standardized code in in the texlang_core::error module to
do the formatting. So I think it's about the same amount of work.

The nice things about this new approach:

- Every error is a different concrete type, so in theory it's possible to
  introspect the error and do something other than just print it.

- The main code is not cluttered with string error messages; instead the
  concrete type is just instantiated.

This commit implements the pattern for the new prefix commands. I think
though we can slowly migrate the rest of the code base to it, and eventually
destroy TokenError and EndOfInputError.