Resurrect the command docs system
The command docs system (in which you can get the documentation for a
particular TeX command) was broken/lost at some point when the control sequence
name interner was created. This commit brings it back.

It also re-pivots the command type, and stores the command ID in a separate
map. This results in a performance improvement, probably due to cache
locality but I'm not 100% sure.