Make the command type smaller
This makes the command type 16 bytes instead of 24 bytes. This is done
by moving the variable address field into the main enum and making it 32 bits
instead of 64 bits. This commit also changes the usizes in the variables
system to u32s.

I think this is actually nice because it's in the spirit of TeX that the
system be independent of the architecture it's running on. The limitation is
that you can now only have 4GB of memory in count registers....which is fine.
Actually the allocation commands still support more memory.