Change grouping map API to accept scope
Grouping map previously had two insert methods: `insert` and `insert_global`.
This resulted in a lot of duplicate code downstream because which method to
invoke was usually determined by a boolean at runtime.

This change adds a new scope enum with two variants (local and global), adds
this as an argument to `insert`, and removes the `insert_global` method. The
same change is made to the commands map.

Interestingly, this refactor already exposed a bug which is that \countdef
does not honor the \global prefix command. This bug was introduced because
\countdef inserts a new command into the commands map without thinking about
the scope. Now that insert explicitly requires a scope, the code looks strange
because it hardcodes the local scope. Thinking about this hardcoding makes
you realize it should support inserting in the global scope, via a \global
prefix.