Remove ArrayVec semi-premature optimization
The macros system originally used ArrayVec to store parameters, which are
guaranteed to be fewer than 9 per macro. This voids a pointer chase from using
Vec. Remove the ArrayVec usages does result in a 3% performance regression,
but I think it is worth it for one less dependency and simpler code.