Move \tracingmacros fully into the stdlib
The \tracingmacros command controls whether the engine should output debug
information during macro expansions. Because macro expansions are handled in
the core/kernel, the initial implementation was inside there. In particular,
the i32 variable that stores the \tracingmacros value was in the base state.
The problem with this is that it makes the variables system more complex
because i32s can be both in the base and custom states.

As a part of removing i32 variables from the base state, this commit moves
macro tracing fully out of the core and into the stdlib. The main ingredient
is to add a "callback" mechanism in the macros system. Whenever a macro is
expanded, this callback is invoked. The callback can be customized entirely,
and is now optionally provided when initializing a VM. The implementation
of \tracingmacros then provides a custom callback that can be provided like
this.

Overall this idea is really nice because it deeply embodies the fundamental
philosophy of the Texcraft project that composition is the primary goal.
Concretely, if another user has a better implementation of \tracingmacros,
they can now easily use it with Texcraft.