Rename scratch space to buffer
I was reading https://github.com/uutils/coreutils/pull/3600 which is a PR that
adds a memory reuse mechanism similar to what I did for TeX macros. I realized
that the right name for the reused memory is "buffer" rather than "scratch
space", which is the name I had used.