New variables API
This commit totally rewrites the Texlang variables API. The initial motivation
was to unlock serializing Texlang VMs when there is an active group (in which
case the restore values need to be serialized somehow). But the new API seems
to be also quite simpler both for adding new variables and for consuming
variable values out of the input stream.

I won't write more because describing how the new API is different from the
old API doesn't seem useful, and the new API is documented quite well (I think)
in the crate.