[Serializable VMs] preliminary map changes
This is the first commit in the project to make the Texlang VM serializable.
With serializable VMs, the complete state of the VM can be serialized to bytes,
deserialized later, and computation can resume where the VM left off.
We may also call this snapshotting, because we can save the current state of
the VM before doing more computation.

Knuth's TeX has an analagous feature which is described in section 1299 of
the TeXBook ("dumping and undumping the tables"). This feature is a little
limited though and is intended only to save the result of consuming format
files like LaTeX. For example, the state of the intepreter can't be saved after
some typesetting has already happened. Texlang VM serialization will be much
more flexible. Moreover, we expect it to be easier to implement
because, unlike Knuth, we don't have to write a custom serialization format -
we'll just use serde.

This commit changes the API of the commands map as the first step of adding
serialization support to the map. Serializing the map is hard because it
contains function pointers which cannot be serialized directly. The idea here,
implemented partially, is to support determining the original control sequence
name for any command currently active in the map. When deserializing we will
provide all of the initial built ins, and use this data to compute the active
commands at the time of serialization.