Put tags inside relevant command variants
It seems that Texcraft's tags system is only applicable to execution and
expansion primitives. This commit moves the tag from being outside the enum
into the two relevant variants. For some reason this doesn't cause the enum
to increase in size - it's still 16 bytes -- which is awesome. I'm not sure
why...

With this change all of the command logic really is inside the enum, and so
the enum is renamed from Fn to Command. What was previously called Command
is renamed to BuiltIn, which I think strongly inidicates its purpose as being
a command provided at initialization. But in the future we could always rename
it to Initial or InitialCommand.