Add VM handlers trait
The VM handlers trait describes how non-execution-command tokens are handled
in the main VM loop. Previously two handlers were passed in as function
pointers. This turns the pointers into a trait, allowing them to be inlined
in the VM loop, and adds a third handler for the unforseen case of an
unexpanded expansion command.