Make ExpandedStream a concrete type
This commit twiddles the type design in streams.rs for the umpteenth
time. The main problem its solving is a better way to be generic over
expanded streams. Previously, this was done with a trait. However the trait
is a bit annoying from an ergonomics perspective. I'm also concerned about
two copies of each function being generated for ExpansionInput and
ExecutionInput and LLVM not deduplicating them.

With the new type design, ExpandedStream is a concrete type and the way
to be generic over expanded streams is to cast either the ExpansionInput
or the ExecutionInput to an ExpandedStream, potentially using the AsMut
implementation. One unexpected benefit of this approach is that a lot of
like the variables API now only works with ExpandedStream rather than
ExpansionInput. This feels right, because why does the variables API care about
the expansion input primitive API.

The commit also dumps the RefVM trait.