Use const arrays for default cat codes (#2)
The default cat codes for ASCII characters in both INITEX and plainTeX are
provided using const arrays of length 128. The existing method that populates
the default in a map uses the plainTeX array. The motivation is that we will
be soon moving the map out of the core crate and want to provide a good
default method for getting the cat code of a character. With this const arrays,
the default will just be an index read on the array.

This commit also updates some docs.