Introduce a parsable trait
I realized that the parsing API (which parses TeX grammar elements) could
be a lot simpler if things we can parse (relations, numbers, catcodes) all
implemented a common trait. Consumers then wouldn't need to know which magic
function in the parsing module to invoke.

Additionally, in the future I see some extra benefits:

- Ability to parse multiple elements of the grammar in one go by implementing
  the trait for tuples of types that implement the trait.

- Some kind of uniform error handling for parsing errors.

This commit adds the trait and implements it for the relation type. At the
same time I realized the relation type is a copy of std::cmp::Ordering, so
I migrated to that type.