Structured errors (commit 1)
This huge commit replaces all of the unstructured anyhow errors in Texlang
with a new structured errors system. With anyhow, errors throughout the code
base could be literally anything. In this new system all errors are either:

- A token error
- An end of input error
- A invalid code path error (an \else without a preceeding \if)
- A propagated error that contains one of the first 3 cases.

When I originally started working on Texcraft using anyhow made sense for
two reasons. First, it allowed me to focus on the happy path and not design
an errors system upfront. Second, my thinking was that because Texcraft allows
arbitrary execution/expansion/varible commands to be written an plugged in,
it should allow arbitrary errors.

The first reason is no longer applicable now that I've been hacking on this
for a while. I now have a much better sense of what kind of errors can
occur, and how they get propagated through the VM.

For the second reason: to have any chance of good error messages it seems
the error type has to have more structure. In particular it seems reasonable
that all errors ultimately point to a specific piece of TeX code, which the
new system enforces.

This is just a merge of the initial work. There is much more to be done,
including:

- writing documentation for the new errors system.

- auditing all existing errors in the code base and checking that the error
  messsage looks good. I already started doing this with the help of a
  new texlang-all-errors binary.

- doing the previous step will require adding a little more to the errors
  system: in particular, notes are currently strings but for the errors in
  the conditionals module we'll want the ability to refer to tokens in the
  notes.

- fixing all the spots where errors are propagated to add the correct
  propagation context.