Add an iterator to grouping map
I added a clone with conversion method to grouping map some time ago, and
have tweaked it more than once to conform to changing requirements. I think
the problem is that clone with conversion is too rigid. Instead, the more
idiomatic Rust approach is to have an iterator which you can mutate and
then collect into a new collection. This commit adds such an iterator.