Serializable VMs (1/N) (#3)
This is the first commit in the serializable VMs project, which is biggish
project to support making the Texcraft VM serializable. The main reason why
this isn't trivial in the usual Rust way is because the VM contains function
pointers (of primitives) and thus you can't trivially serde the VM.

This commit adds (or fixes) some of the initial infrastructure that was
added (e.g., the command key type) and adds support for serializing primitives
only. Support for other commands will come later. One of the main additions
here is a unit testing utility for easy unit testing serding stuff.