Support serding with message pack (#3)
One bug with this (de)serializer is that the registers array overflows the
stack, I'm guessing becuase the call stack gets bigger than with serde_json.
I put the array behind a Box which is unfortunate because there's a small
runtime cost to that. Maybe in the future there will be a better solution.

I also put all VMs behind boxes. I thought this may have solved the stack
overflow but it didn't. Nonetheless it's worth doing because the VM can be
big.

After this commit all serde unit tests run both with json and message
pack.