Support serializing in an active group (#3)
To do this we need to support (de)serializing the save stack. This is a bit
of a pain as the save stack contains function pointers that need to be
cross-referenced with what's in the command map. In the end the code touches
the commands map, the variables API, and the serde module, so there are
more pub(crate)s than I would like. I tried to refactor things to minimize
cross module deps.

But it works!