Implement active characters
An active charactor is a charactor with category code 13. It behaves
essentially like a control sequence. This commit implements active characters
such they can now be used wherever control sequences can be used (\def, \let
\countdef etc.). The only exception is we still only support initializing the
commands map with control sequences, and documentation support is not 100%
there.

The implementation is actually pretty easy because of Rust's type system.
Everywhere we previously accepted a CsName we just accept a new
token::CommandRef type, which is an enum with one variant for control
sequences and one variable for active characters. Fixing the resulting
type errors essentially implements the feature.