Tweak the expansion command API
In the original expansion command API, an expansion command returns a vector
of tokens which are put at the front of the input. This seemed to make sense
as my naive view of expansion commands was that they consume input tokens
and push output tokens. The reality is more complicated:

- Most expansion commands don't actually push tokens, but instead mutate
  the input token stream in some other way. E.g., \input starts a new file,
  \endinput ends the current file, \expandafter triggers other expansions
  but doesn't itself return any expanded tokens.

- Returning a vector is inefficient because it means either an allocation
  or grabbing one of the token buffers. The more efficient alternative is
  to push tokens directly onto the token stack.

Given these considerations, this commit changes the API so that expansion
commands just return (), like execution commands. The expectation is that
all mutations occur via methods on the ExpansionInput type.

The digits of pi benchmark improves with this commit which is somewhat
surprising. Perhaps the generated code in streams.rs is simpler now that
the vector that is returned doesn't need to be inspected.