Implement `\toks` and `\toksdef`
This involved implementing the first new variable type in Texcraft in years:
the token list.

This commit contains an interesting optimization around the use of vectors.
A token list is a vector of tokens and unlike existing variable types (like
integers) token lists incur a performance cost to create. In this commit we
fully recycle all vectors used in token list variables. During a VM run
at most (number of token list variables + number of token list save stack
entries + 1) vectors will be used. This is achieved by recycling vectors
at the right point. E.g., if we are overwriting the value of a token list,
we take the existing vector and put it back into the token buffers pool.