Fix bug in undelimited macro argument parsing
When parsing an undelimited argument for a macro, spaces must first be trimmed
from the unexpanded input stream. This bug was found by re-enabling a unit
test for exercise 7 in the TeXBook chapter 20.

There was previously a bug in the lexer that hid this bug. The bug in the lexer
was that spaces were trimmed after parsing a control sequence like \+, but this
is not right. When that bug was fixed as part of the recent lexer re-write,
the bug this commit fixes appeared. However because the unit test was disabled,
I didn't catch it. But also there were certain cases the unit test didn't
catch, for example the second undelimited argument must also have its spaces
trimmed.