Remove some things from the VM
Before this commit, the VM contained some things (like access to the file
system) that weren't used in the VM. I'm trying to implement recoverable
errors and some of these things are causing issues with serding. Thinking about
the problem, it seems like the VM should not include anything that the VM
doesn't need. The VM is already bloated and global enough.

Additionally, the TexlangState trait should only include hooks used within
the VM.

With this in mind, this commit creates a new crate texlang-common that contains
stuff that could be in the VM because it is universally needed (like logging)
but that doesn't need to be in the VM.