Recoverable errors
Most errors in TeX, like an index out of bounds error in \count, are
recoverable. In these cases TeX has a default behaviour - e.g., in the
case of \count, the index is assumed to be 0).

This commit adds infrastructure for recoverable errors. The main
addition is a new method on Texlang state which functions must
call when an error occurs. This determines whether the fail, or to
recover.