TeX font metric deserializer
I started working on parsing TFM and PL files nearly 18 months ago, while
travelling from New York to Hawaii. The branch has been around since then, but
there is/was a lot of work to finish it.

This commit is a polishing/rewrite of one part of that work, which is reading
TFM files. This part is actually needed for parsing plain.tex, so it's worth
committing now. I can next start to think about integrating fonts with
Texlang.