tfm: add initial PL parsing
This commit adds the code that parses a PL AST into a struct with all the
data. There are still a few AST nodes left to handle (extensible characters
and next larger character chains).

A follow-up commit will then use this code to re-implement PLtoTF by adding
a converter from `pl::File` to `crate::File` and then a serializer for
`crate::File`.