Add first version of tftopl binarys
This code has been sitting on my computer for 2 months+ and I want to merge
it into mainline so that I don't forget about it. The code is good, but does
not necessarily form a perfectly meaningful checkpoint. It also has a lot of
TODOs!

The main work here is implementing the tftopl binary. This involves both
adding the binary itself (built using Clap), but also changing the tfm crate
to support outputting .pl files. This is done by adding various "lowering"
methods to different trees/data structures.

The next step is to get a large corpus of .tfm files, run both Texcraft's
tftopl and Knuth's tftopl over them, and then find all the differences.
These differences are bugs. Eventually the new binary should return the
same information as Knuth's. A large corpus can be obtained from the TeXLive
repo.