Add `HasDefaultBuiltInCommands` trait
This trait is implemented for state types, and provides a default set of
built-in commands for that state. When this is trait is implemented
deserialization is much easier (and idiomatic) because the built-in commands
don't need to be explicitly provided and can be taken from the trait.
Additionally, this trait allows for a simple VM `new` constructor to be
added.

This commit also updates the documentation accordingly.

Finally in this commit I standardize the entire codebase and docs on the
phrase "built-in commands" for the collection of commands that are in the VM
when it is initialized. Previously there was a mish mash of different phrases
(initial commands, initial primitives, built-in primitives, etc.).

With this naming convention we understand that a "primitive" is a specific
kind of command (execution, expansion, variable) that is implemented in Rust
code (as opposed to e.g. a TeX macro). While most/all built-in commands will
be primitivies, in theory it's possible to provide a TeX user-defined macro.
This is why I used the phrase "built-in command".