tfm: add a KernIndex lig/kern op
In .tfm files, lig/kern kern instructions do not contain the kern itself,
but an index into a kerns array. The previous code piggybacked on the Kern
variant of lig/kern op to hold this, but this is really bad as it means the
type system can no longer tell you whether an entry is an index or a kern.
This commit adds a new variant for this.