tftopl: add fuzz testing
So far I have mostly done diff testing for tftopl using the corpus of .tfm
files from CTAN. The problem with this corpus is that the files are too
structured: they are almost all the output of the same program (pltotf)
and so don't really exercise as many edge cases as one would like.

This commit adds fuzz testing, which in a sense is on the opposite end of
the spectum: the .tfm files generated are random and structureless.

Already the fuzz tests have caught one bug (a simple one to be fair). This
is fixed here.