tfm: change .pl extension to .plst
This commit changes the extension of all property list files in the repo from
.pl to .plst. It also changes the default extension used for these files to
.plst.

The main motivation is that after submitting all of these files, GitHub now
marks the project as being mostly developed in Perl and a little in Prolog.
Of course this is just a "marketing" problem. However I think it's good to
advertise up front that the project is written in Rust for a number of
reasons. Ultimately, people are more interested in Rust projects these days.

Additionally, I essentially failed to do pltotf diff testing on CTAN because
of the extension collision. The diff testing was mostly false postives because
so many .pl files are not property list files, but rather Perl scripts. Going