tfm: documentation and general polishing
Based on the fuzzing results, it seems this crate's implementation of tftopl
may be complete! This is the first in a series of commits that is adding
docs and in general just polishing the crate. The overall goal is to make
the API documentation (i.e. rustdocs) easy to navigate.