pltotf: mimic error handling
This (huge) commit adds the first code that tries to mimic the error handling
in pltotf. The first run of the fuzzer highlighted diffs in stderr, and then
I realized that the E2E tests weren't verifying stderr for pltotf tests.

The main thing this commit adds is a test case all-errors.plst that will
eventually contain almost all errors that pltotf can raise. I've then changed
the code so that about ~20 errors kinds are handled correctly. This mean
we're perhaps halfway there, but I wanted to commit this now becuase the
commit was getting too big.