tfm: refactor CST parsing
This refactor dramatically simplifies CST parsing. It does so, in part, by
removing the lexer, which turned out to be a bad abstraction. Much of the
complexity of the old parsing code was because chars had to be mapped to
tokens and then tokens to CST elements, but tokens didn't really represent
anything meaningful in the PL grammar. The mapping code was then complicated,
especially the token to CST step. With this commit the CST is parsed in
one pass straight from the PL string.

This commit also handles CRLF line endings correctly.