typesetting: initial scaffolding
This commit contains some initial scaffolding for the typesetting crate(s)
that will implement Knuth/TeX's typesetting engine. There is one Texcraft-
agnostic crate (so it can be used outside of the Texcraft project) and another
crate that provides Texlang bindings for the first create. The initial
scaffolding has some basic implementation code, but I'm mostly committing
so the traits/interfaces are in.