Error system: simpler and more correct
This commit is a nontrivial refactor of the errors that makes it both simpler
to use and less likely to have bugs.

I was led to this commit after realizing that the recoverable error system
was tricky to use correctly. In TeX, every error can be gracefully handled
and processing can continue. When an error occurs, the Texlang code is
supposed to check (using the recoverable error hook) whether to gracefully
handle the error and keep going, or fail and propagate the error back to the
VM.

There were two problems in practice. Problem 1 is that in many cases the
recoverable error hook isn't invoked so all errors are fatal errors.

Problem 2 is that it's not clear where exactly the recoverable error hook
should be invoked and based on this it may accidentally get invoked for the
same error twice. For example, when the \def command starts running it looks
for a control sequence to define. Where should the missing control sequence
error be handled and recovered from? Should it happen in the parse code, or
in the \def command?

Problem 1 and 2 can be summarized by saying that for each error, the
hook should be invoked once and only once.

I realized the Rust type system can actually enforce this! The idea is that
the hook takes in a Rust type A (which the user constructed, in this commit an
implementation of TexError) and returns a Rust type B (created by the VM,
user can't do anything about it, type Error in this commit).
Type B is the only thing that can be propagated and can't be retried. Thus
when writing Texlang code if you want to create an error you are forced to
go through the hook.

When I say "go through the hook" I actually mean invoking a new method
error() on the VM.

Once that was designed, I realized I could also simplify the source code
tracing situation. Before this commit error handling code was cluttered with
many calls to the tracer. But with this new setup, the error() method can
automatically perform all the tracing by inspecting the TexError. The code
writer only needs to provide the tokens that are implicated. This makes
creating errors much more simple.